add split plugin files into separate folders
