+++ javascript/modules/client/src/client-addressList/index.ts
@@ -1,0 +1,1 @@
+export * from './client-addressList'
