+++ javascript/modules/client/src/client-addressList/internal/client-addressList/estimation.ts
@@ -1,25 +1,26 @@
 import { Random } from "@aragon/sdk-common";
-import { ContextPlugin } from "../../context-plugin";
-import { ClientCore } from "../core";
-import { GasFeeEstimation } from "../interfaces/common";
+import { ClientCore } from "../../../internal/core";
+import { GasFeeEstimation } from "../../../internal/interfaces/common";
 import {
-  IClientErc20Estimation,
+  ContextPlugin,
   ICreateProposalParams,
   IExecuteProposalParams,
   IVoteProposalParams,
-} from "../interfaces/plugins";
+} from "../../../plugin-common";
+import { IClientAddressListEstimation } from "../interfaces/client";
 
-export class ClientErc20Estimation extends ClientCore
-  implements IClientErc20Estimation {
+export class ClientAddressListEstimation extends ClientCore
+  implements IClientAddressListEstimation {
   constructor(context: ContextPlugin) {
     super(context);
   }
+
   /**
    * Estimates the gas fee of creating a proposal on the plugin
    *
    * @param {ICreateProposalParams} _params
    * @return {*}  {Promise<GasFeeEstimation>}
-   * @memberof ClientErc20Estimation
+   * @memberof ClientAddressListEstimation
    */
   public createProposal(
     _params: ICreateProposalParams,
@@ -30,33 +31,18 @@ export class ClientErc20Estimation extends ClientCore
     } else if (!signer.provider) {
       throw new Error("A web3 provider is needed");
     }
-
-    // TODO: Remove below as the new contracts are ready
-
+    // TODO: Implement
     return Promise.resolve(
       this.web3.getApproximateGasFee(Random.getBigInt(BigInt(1500))),
     );
-
-    // TODO: Uncomment below as the new contracts are ready
-    /*
-    const erc20VotingInstance = ERC20Voting__factory.connect(
-      this._pluginAddress,
-      signer
-    );
-
-    return erc20VotingInstance.estimateGas.newVote(
-      ...unwrapProposalParams(params),
-    ).then((gasLimit) => {
-      return this.web3.getApproximateGasFee(gasLimit.toBigInt());
-    });
-    */
   }
+
   /**
    * Estimates the gas fee of casting a vote on a proposal
    *
    * @param {IVoteProposalParams} _params
    * @return {*}  {Promise<GasFeeEstimation>}
-   * @memberof ClientErc20Estimation
+   * @memberof ClientAddressListEstimation
    */
   public voteProposal(_params: IVoteProposalParams): Promise<GasFeeEstimation> {
     const signer = this.web3.getConnectedSigner();
@@ -65,18 +51,18 @@ export class ClientErc20Estimation extends ClientCore
     } else if (!signer.provider) {
       throw new Error("A web3 provider is needed");
     }
-    // TODO: remove this
+    // TODO: Implement
     return Promise.resolve(
       this.web3.getApproximateGasFee(Random.getBigInt(BigInt(1500))),
     );
   }
 
   /**
-   * Estimates the gas fee of executing an ERC20 proposal
+   * Estimates the gas fee of executing an AddressList proposal
    *
    * @param {IExecuteProposalParams} _params
    * @return {*}  {Promise<GasFeeEstimation>}
-   * @memberof ClientErc20Estimation
+   * @memberof ClientAddressListEstimation
    */
   public executeProposal(
     _params: IExecuteProposalParams,
@@ -87,7 +73,7 @@ export class ClientErc20Estimation extends ClientCore
     } else if (!signer.provider) {
       throw new Error("A web3 provider is needed");
     }
-    // TODO: remove this
+    // TODO: Implement
     return Promise.resolve(
       this.web3.getApproximateGasFee(Random.getBigInt(BigInt(1500))),
     );
