+++ javascript/modules/client/src/client-addressList/internal/constant.ts
@@ -1,0 +1,3 @@
+// NOTE: This address needs to be set when the plugin has been published and the ID is known
+export const ADDRESSLIST_PLUGIN_ID =
+  "0x1234567890123456789012345678901234567890";
