+++ javascript/modules/client/src/client-addressList/internal/graphql-queries/settings.ts
@@ -1,0 +1,11 @@
+import { gql } from "graphql-request";
+
+export const QueryAddressListPluginSettings = gql`
+query AddressListPluginSettings($address: ID!) {
+  allowlistPackage(id: $address){
+    supportRequiredPct
+    participationRequiredPct
+    minDuration
+  }
+}
+`;
