+++ javascript/modules/client/src/client-addressList/internal/interfaces/client.ts
@@ -1,0 +1,126 @@
+// This file contains the definitions of the AddressList DAO client
+
+import { IClientCore } from "../../../internal/interfaces/core";
+import {
+  DaoAction,
+  GasFeeEstimation,
+  IInterfaceParams,
+} from "../../../internal/interfaces/common";
+import { BigNumber } from "@ethersproject/bignumber";
+import {
+  ExecuteProposalStepValue,
+  ICanVoteParams,
+  ICreateProposalParams,
+  IExecuteProposalParams,
+  IPluginSettings,
+  IProposalQueryParams,
+  IProposalSettings,
+  IVoteProposalParams,
+  ProposalBase,
+  ProposalCreationStepValue,
+  ProposalListItemBase,
+  SubgraphAction,
+  SubgraphProposalBase,
+  SubgraphVoterListItemBase,
+  VoteProposalStepValue,
+  VoteValues,
+} from "../../../plugin-common/interfaces";
+
+// Address List
+export interface IClientAddressListMethods extends IClientCore {
+  createProposal: (
+    params: ICreateProposalParams,
+  ) => AsyncGenerator<ProposalCreationStepValue>;
+  voteProposal: (
+    params: IVoteProposalParams,
+  ) => AsyncGenerator<VoteProposalStepValue>;
+  executeProposal: (
+    params: IExecuteProposalParams,
+  ) => AsyncGenerator<ExecuteProposalStepValue>;
+  canVote: (params: ICanVoteParams) => Promise<boolean>;
+  getMembers: (addressOrEns: string) => Promise<string[]>;
+  getProposal: (propoosalId: string) => Promise<AddressListProposal | null>;
+  getProposals: (
+    params: IProposalQueryParams,
+  ) => Promise<AddressListProposalListItem[]>;
+  getSettings: (pluginAddress: string) => Promise<IPluginSettings | null>;
+}
+
+export interface IClientAddressListEncoding extends IClientCore {
+  updatePluginSettingsAction: (
+    pluginAddress: string,
+    params: IPluginSettings,
+  ) => DaoAction;
+  addMembersAction: (
+    pluginAddress: string,
+    members: string[],
+  ) => DaoAction;
+  removeMembersAction: (
+    pluginAddress: string,
+    members: string[],
+  ) => DaoAction;
+}
+export interface IClientAddressListDecoding extends IClientCore {
+  updatePluginSettingsAction: (data: Uint8Array) => IPluginSettings;
+  addMembersAction: (data: Uint8Array) => string[];
+  removeMembersAction: (data: Uint8Array) => string[];
+  findInterface: (data: Uint8Array) => IInterfaceParams | null;
+}
+export interface IClientAddressListEstimation extends IClientCore {
+  createProposal: (
+    params: ICreateProposalParams,
+  ) => Promise<GasFeeEstimation>;
+  voteProposal: (params: IVoteProposalParams) => Promise<GasFeeEstimation>;
+  executeProposal: (
+    params: IExecuteProposalParams,
+  ) => Promise<GasFeeEstimation>;
+}
+/** Defines the shape of the AddressList client class */
+export interface IClientAddressList {
+  methods: IClientAddressListMethods;
+  encoding: IClientAddressListEncoding;
+  decoding: IClientAddressListDecoding;
+  estimation: IClientAddressListEstimation;
+}
+
+export type AddressListProposal = ProposalBase & {
+  result: AddressListProposalResult;
+  settings: IProposalSettings;
+  votes: Array<{ address: string; vote: VoteValues }>;
+  totalVotingWeight: number;
+};
+
+export type AddressListProposalListItem = ProposalListItemBase & {
+  result: AddressListProposalResult;
+};
+export type AddressListProposalResult = {
+  yes: number;
+  no: number;
+  abstain: number;
+};
+
+export type SubgraphAddressListVoterListItem = SubgraphVoterListItemBase;
+
+export type SubgraphAddressListProposalListItem = SubgraphProposalBase;
+export type SubgraphAddressListProposal = SubgraphProposalBase & {
+  createdAt: string;
+  actions: SubgraphAction[];
+  supportRequiredPct: string;
+  participationRequired: string;
+  voters: SubgraphAddressListVoterListItem[];
+  votingPower: string;
+};
+
+export type ContractAddressListInitParams = [
+  string,
+  string,
+  BigNumber,
+  BigNumber,
+  BigNumber,
+  string[],
+];
+
+export type IAddressListPluginInstall = {
+  addresses: string[];
+  settings: IPluginSettings;
+};
