+++ javascript/modules/client/src/client-addressList/internal/utils.ts
@@ -1,0 +1,125 @@
+import { hexToBytes, strip0x } from "@aragon/sdk-common";
+import { DaoAction } from "../../internal/interfaces/common";
+import {
+  ProposalMetadata,
+  SubgraphAction,
+  SubgraphVoteValuesMap,
+  VoteValues,
+} from "../../plugin-common";
+import { computeProposalStatus } from "../../plugin-common/utils";
+import {
+  AddressListProposal,
+  AddressListProposalListItem,
+  SubgraphAddressListProposal,
+  SubgraphAddressListProposalListItem,
+  SubgraphAddressListVoterListItem,
+} from "./interfaces/client";
+
+export function toAddressListProposal(
+  proposal: SubgraphAddressListProposal,
+  metadata: ProposalMetadata,
+): AddressListProposal {
+  const startDate = new Date(
+    parseInt(proposal.startDate) * 1000,
+  );
+  const endDate = new Date(parseInt(proposal.endDate) * 1000);
+  const creationDate = new Date(
+    parseInt(proposal.createdAt) * 1000,
+  );
+  return {
+    id: proposal.id,
+    dao: {
+      address: proposal.dao.id,
+      name: proposal.dao.name,
+    },
+    creatorAddress: proposal.creator,
+    metadata: {
+      title: metadata.title,
+      summary: metadata.summary,
+      description: metadata.description,
+      resources: metadata.resources,
+      media: metadata.media,
+    },
+    startDate,
+    endDate,
+    creationDate,
+    actions: proposal.actions.map(
+      (action: SubgraphAction): DaoAction => {
+        return {
+          data: hexToBytes(strip0x(action.data)),
+          to: action.to,
+          value: BigInt(action.value),
+        };
+      },
+    ),
+    status: computeProposalStatus(proposal),
+    result: {
+      yes: proposal.yes ? parseInt(proposal.yes) : 0,
+      no: proposal.no ? parseInt(proposal.no) : 0,
+      abstain: proposal.abstain ? parseInt(proposal.abstain) : 0,
+    },
+    settings: {
+      // TODO
+      // this should be decoded using the number of decimals that we want
+      // right now the encoders/recoders use 2 digit precission but the actual
+      // subgraph values are 18 digits precision. Uncomment below for 2 digits
+      // precision
+
+      // minSupport: decodeRatio(
+      //   BigInt(proposal.supportRequiredPct),
+      //   2,
+      // ),
+      // minTurnout: decodeRatio(
+      //   BigInt(proposal.participationRequiredPct),
+      //   2,
+      // ),
+      // TODO DELETE ME
+      minSupport: parseFloat(
+        proposal.supportRequiredPct,
+      ),
+      minTurnout: parseFloat(
+        proposal.participationRequired,
+      ),
+      duration: parseInt(proposal.endDate) -
+        parseInt(proposal.startDate),
+    },
+    totalVotingWeight: parseInt(proposal.votingPower),
+    votes: proposal.voters.map(
+      (voter: SubgraphAddressListVoterListItem) => {
+        return {
+          address: voter.voter.id,
+          vote: SubgraphVoteValuesMap.get(voter.vote) as VoteValues,
+        };
+      },
+    ),
+  };
+}
+export function toAddressListProposalListItem(
+  proposal: SubgraphAddressListProposalListItem,
+  metadata: ProposalMetadata,
+): AddressListProposalListItem {
+  const startDate = new Date(
+    parseInt(proposal.startDate) * 1000,
+  );
+  const endDate = new Date(parseInt(proposal.endDate) * 1000);
+  return {
+    id: proposal.id,
+    dao: {
+      address: proposal.dao.id,
+      name: proposal.dao.name,
+    },
+    creatorAddress: proposal.creator,
+    metadata: {
+      title: metadata.title,
+      summary: metadata.summary,
+    },
+    startDate,
+    endDate,
+    status: computeProposalStatus(proposal),
+    result: {
+      yes: proposal.yes ? parseInt(proposal.yes) : 0,
+      no: proposal.no ? parseInt(proposal.no) : 0,
+      abstain: proposal.abstain ? parseInt(proposal.abstain) : 0,
+    },
+  };
+}
