+++ javascript/modules/client/src/client-erc20/index.ts
@@ -1,0 +1,1 @@
+export * from "./client-erc20"
