+++ javascript/modules/client/src/client-erc20/internal/constants.ts
@@ -1,0 +1,2 @@
+// NOTE: This address needs to be set when the plugin has been published and the ID is known
+export const ERC20_PLUGIN_ID = "0x1234567890123456789012345678901234567890";
