+++ javascript/modules/client/test/integration/client-addressList/encoding.test.ts
@@ -6,7 +6,7 @@ import { ClientAddressList, Context, ContextPlugin } from "../../../src";
 import {
   IAddressListPluginInstall,
   IPluginSettings,
-} from "../../../src/internal/interfaces/plugins";
+} from "../../../src";
 import { bytesToHex, InvalidAddressError } from "@aragon/sdk-common";
 import { contextParamsLocalChain } from "../constants";
 describe("Client Address List", () => {
