+++ javascript/modules/client/test/integration/client-erc20/estimation.test.ts
@@ -1,14 +1,8 @@
 // @ts-ignore
 declare const describe, it, expect, beforeAll, afterAll;
 
-import { ClientErc20, Context, ContextPlugin } from "../../../src";
+import { ClientErc20, Context, ContextPlugin, ICreateProposalParams, IExecuteProposalParams, IVoteProposalParams, VoteValues } from "../../../src";
 
-import {
-  ICreateProposalParams,
-  IExecuteProposalParams,
-  IVoteProposalParams,
-  VoteValues,
-} from "../../../src/internal/interfaces/plugins";
 import { contextParamsLocalChain } from "../constants";
 import * as ganacheSetup from "../../../../../helpers/ganache-setup";
 import * as deployContracts from "../../../../../helpers/deployContracts";
