+++ javascript/modules/client/test/integration/client-erc20/methods.test.ts
@@ -1,29 +1,10 @@
 // @ts-ignore
 declare const describe, it, beforeAll, afterAll, expect, test;
 
-import {
-  Client,
-  ClientErc20,
-  Context,
-  ContextPlugin,
-  SortDirection,
-} from "../../../src";
+import { Client, ClientErc20, Context, ContextPlugin, ExecuteProposalStep, ICanVoteParams, ICreateProposalParams, IExecuteProposalParams, IProposalQueryParams, IVoteProposalParams, ProposalCreationSteps, ProposalSortBy, ProposalStatus, SortDirection, VoteProposalStep, VoteValues } from "../../../src";
 import * as ganacheSetup from "../../../../../helpers/ganache-setup";
 import * as deployContracts from "../../../../../helpers/deployContracts";
 
-import {
-  ExecuteProposalStep,
-  ICanVoteParams,
-  ICreateProposalParams,
-  IExecuteProposalParams,
-  IProposalQueryParams,
-  IVoteProposalParams,
-  ProposalCreationSteps,
-  ProposalSortBy,
-  ProposalStatus,
-  VoteProposalStep,
-  VoteValues,
-} from "../../../src/internal/interfaces/plugins";
 import { InvalidAddressOrEnsError } from "@aragon/sdk-common";
 import {
   contextParams,
