+++ javascript/modules/client/test/integration/client/decoding.test.ts
@@ -11,7 +11,7 @@ import {
   IRevokePermissionParams,
   IWithdrawParams,
   Permissions,
-} from "../../../src/internal/interfaces/client";
+} from "../../../src";
 import { contextParamsLocalChain } from "../constants";
 import { keccak256 } from "@ethersproject/keccak256";
 import { toUtf8Bytes } from "@ethersproject/strings";
